SourceCode/CareerButterflySystem/one_stop_shop/requirements.txt
streamlit==1.32.2
python-docx==1.1.0
asyncio==3.4.3
nltk==3.8.1
PyPDF2==3.0.1
openai==1.14.1
streamlit-chat==0.1.1
urllib3==2.2.1